Advance de inventarios

+++ src/controllers/receta.controllers.js
@@ -2,12 +2,18 @@ const catchError = require('../utils/catchError');
 const Recetas = require('../models/Receta');
 const Recet_articulo = require('../models/Recetas_ingredientes')
 const ingrediente = require('../models/Articulo')
+const User = require('../models/User')
 
 
 const getAll = catchError(async(req, res) => {
    
-    console.log("Recta controllers", req.user)
-    const Resulst = await Recetas.findAll()
+    //const 
+    const Resulst = await Recetas.findAll({where:{},
+        include:{
+            model:Recet_articulo,
+            attributes:{exclude:['id','createdAt','updatedAt','recetumId','articuloId']}}},
+        
+    )
     res.status(200).json(Resulst)
 
 });
@@ -15,8 +21,16 @@ const getAll = catchError(async(req, res) => {
 
 const Create = catchError(async(req, res)=>{
      
-    const {name, T_preperacion, T_coccion, imagen , video} = req.body
-    const results = await Recetas.create(req.body)
+    const {nombre, T_preparacion, T_coccion, imagen , video} = req.body
+    console.log(req.body)
+    const Recet ={
+        nombre,
+        T_preparacion,
+        T_coccion,
+        imagen,
+        video
+    }
+    const results = await Recetas.create(Recet)
 
     if(!results) return res.status(404).json({"Data":"Don't create receta"})
 
@@ -28,19 +42,19 @@ const Create = catchError(async(req, res)=>{
 const AddIngrediente = catchError(async(req, res)=>{
     const id = parseInt(req.params.id)
     console.log(id)
-    const {recetumId, ArticuloId, cantidad, unidad } = req.body
+    const {recetumId, articuloId, cantidad, unidad } = req.body
 
-    const isValue = await Recet_articulo.findOne({where:{recetumId:id, ArticuloId }})
+    const isValue = await Recet_articulo.findOne({where:{recetumId:id, articuloId }})
   console.log("Este is value",isValue)
    if(!isValue){
        if(id){
            const articulo = await ingrediente.findOne({
-               where:{id:ArticuloId}
+               where:{id:articuloId}
        })        
            const datos={
                 items:articulo.items,
                 recetumId:id,
-                ArticuloId,
+                articuloId,
                 cantidad, 
                 unidad
            }
@@ -59,10 +73,10 @@ const getOne = catchError(async(req, res)=>{
     const result = await Recetas.findOne({where:{id},
         include:[{  
             model:Recet_articulo, 
-            attributes:{exclude:['id','createdAt','updatedAt','recetumId','ArticuloId']}}],       
+            attributes:{exclude:['id','createdAt','updatedAt','recetumId','articuloId']}}],       
         });
-if(!result) return res.status(404).json("no encontrado")
-    return res.status(201).json(result)
+        if(!result) return res.status(404).json("no encontrado")
+        return res.status(201).json(result)
 })
 
 const Update = catchError(async(req, res)=>{
@@ -75,9 +89,13 @@ const Update = catchError(async(req, res)=>{
 })
 
 const Delete = catchError(async(req, res)=>{
-    const id =  parseInt(req.params.id);
-   const results =  await Recetas.destroy({where:{id}}) 
-   res.json({"Data":"Delete sucessfull Receta"})
+    const user = await User.findAll({where:{email:req.user.email}})
+    if(user.tipo === "admin"){
+        const id =  parseInt(req.params.id);
+       const results =  await Recetas.destroy({where:{id}}) 
+       res.json({"Data":"Delete sucessfull Receta"})
+    }
+    res.status(404).json({"Data":"Usuario no Autorizado"})
 })
 
 module.exports = {
